+++ Editor/Playa/Renderer/NativePropertyRenderer.cs
@@ -1,4 +1,6 @@
-﻿using UnityEditor;
+﻿using System.Reflection;
+using SaintsField.Playa;
+using UnityEditor;
 using UnityEngine;
 #if UNITY_2021_3_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
 using System.Linq;
@@ -10,14 +12,22 @@ namespace SaintsField.Editor.Playa.Renderer
 {
     public class NativePropertyRenderer: AbsRenderer
     {
+        private readonly bool _renderField;
+
         public NativePropertyRenderer(SerializedObject serializedObject, SaintsFieldWithInfo fieldWithInfo) : base(fieldWithInfo)
         {
+            _renderField = fieldWithInfo.PropertyInfo.GetCustomAttribute<ShowInInspectorAttribute>() != null;
         }
 #if UNITY_2021_3_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
         private VisualElement _fieldElement;
         // private bool _callUpdate;
         protected override (VisualElement target, bool needUpdate) CreateTargetUIToolkit()
         {
+            if (!_renderField)
+            {
+                return (null, false);
+            }
+
             object value = FieldWithInfo.PropertyInfo.GetValue(FieldWithInfo.Target);
 
             VisualElement container = new VisualElement
@@ -40,6 +50,11 @@ namespace SaintsField.Editor.Playa.Renderer
         // private static void WatchValueChanged(SaintsFieldWithInfo fieldWithInfo,  VisualElement container, bool callUpdate)
         {
             PreCheckResult preCheckResult = base.OnUpdateUIToolKit();
+            if (!_renderField)
+            {
+                return preCheckResult;
+            }
+
             object userData = _fieldElement.userData;
             object value = FieldWithInfo.PropertyInfo.GetValue(FieldWithInfo.Target);
 
@@ -55,7 +70,8 @@ namespace SaintsField.Editor.Playa.Renderer
                 StyleEnum<DisplayStyle> displayStyle = child.style.display;
                 _fieldElement.Clear();
                 _fieldElement.userData = value;
-                _fieldElement.Add(child = UIToolkitLayout(value, ObjectNames.NicifyVariableName(FieldWithInfo.PropertyInfo.Name)));
+                _fieldElement.Add(child = UIToolkitLayout(value,
+                    ObjectNames.NicifyVariableName(FieldWithInfo.PropertyInfo.Name)));
                 child.style.display = displayStyle;
             }
 
@@ -69,6 +85,11 @@ namespace SaintsField.Editor.Playa.Renderer
 
         protected override void RenderTargetIMGUI(PreCheckResult preCheckResult)
         {
+            if (!_renderField)
+            {
+                return;
+            }
+
             object value = FieldWithInfo.PropertyInfo.GetValue(FieldWithInfo.Target);
             FieldLayout(value, ObjectNames.NicifyVariableName(FieldWithInfo
                 .PropertyInfo.Name));
@@ -76,11 +97,20 @@ namespace SaintsField.Editor.Playa.Renderer
 
         protected override float GetFieldHeightIMGUI(float width, PreCheckResult preCheckResult)
         {
+            if (!_renderField)
+            {
+                return 0f;
+            }
             return FieldHeight(FieldWithInfo.PropertyInfo.GetValue(FieldWithInfo.Target), ObjectNames.NicifyVariableName(FieldWithInfo.PropertyInfo.Name));
         }
 
         protected override void RenderPositionTarget(Rect position, PreCheckResult preCheckResult)
         {
+            if (!_renderField)
+            {
+                return;
+            }
+
             object value = FieldWithInfo.PropertyInfo.GetValue(FieldWithInfo.Target);
             FieldPosition(position, value, ObjectNames.NicifyVariableName(FieldWithInfo
                 .PropertyInfo.Name));