+++ Samples~/Scripts/IssueAndTesting/Issue/RequiredOnList.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue
+{
+    public class RequiredOnList: MonoBehaviour
+    {
+        [Required] public Scriptable[] scriptableArray;
+    }
+}