+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.3.3 ##
+
+1.  UI Toolkit now can fall back to an IMGUI custom drawer if the target field is specified to be drawn by an IMGUI drawer (note: `<icon>` in `RichLabel` will not work and will get removed).
+2.  Change the order of static/readonly field of `ShowInInspector` so it can stay at the position where it declared. Change the order of `property` to be above `method`.
+3.  Fix compile error on old Unity version.
+4.  `ColorToggle` supports `Graphic` (`text`, `TMP_Text`, `Image` etc). [#83](https://github.com/TylerTemp/SaintsField/issues/83)
+
+## 3.3.2 ##
+
+1.  Fix `Required` not work on every element of list/array.
+2.  Fix `RichLabel` etc. failed to find a correct fallback drawer like `UnityEvent`
+
 ## 3.3.1 ##
 
 Fix `Dropdown` & `AdvancedDropdown` not work on list/array.